Extract shared schedule formatting funcs used by automation ui components
Avoid duplicate logic by reusing these functions
