+++ src/client/components/automations/AutomationCard.tsx
@@ -1,11 +1,10 @@
 // Individual automation card for automations page gallery
 
-import React from 'react';
 import { useTranslation } from 'react-i18next';
-import type { TFunction } from 'i18next';
-import i18n from '../../i18n';
 import { ChevronRight, Clock, Calendar, AlertCircle } from 'lucide-react';
 import type { AutomationInfo, AutomationPendingConfirmation } from '../../types/automations';
+import { formatTime, formatDayOfWeek, formatDayOfMonth, formatNextRun } from './utils';
+import type { TFunction } from 'i18next';
 
 interface AutomationCardProps {
     automation: AutomationInfo;
@@ -35,80 +34,22 @@ function formatSchedule(automation: AutomationInfo, t: TFunction): string {
     const dayOfWeek = parts[4] ?? '*';
 
     const hourNum = parseInt(hour, 10);
-    const formatTime = (h: number, m: string) => {
-        const am = t('automations.timePeriodAM');
-        const pm = t('automations.timePeriodPM');
-        // If locale has no AM/PM strings, use 24-hour format
-        if (!am && !pm) return `${h.toString().padStart(2, '0')}:${m.padStart(2, '0')}`;
-        const h12 = h % 12 || 12;
-        return `${h12}:${m.padStart(2, '0')} ${h < 12 ? am : pm}`;
-    };
-    const timeStr = formatTime(hourNum, minute);
+    const timeStr = formatTime(hourNum, minute, t);
 
-    // Hourly
     if (hour === '*') {
         return t('automations.hourlyAt', { minute: minute.padStart(2, '0') });
     }
-
-    // Daily
     if (dayOfMonth === '*' && dayOfWeek === '*') {
         return t('automations.dailyAt', { time: timeStr });
     }
-
-    // Weekly
     if (dayOfMonth === '*' && dayOfWeek !== '*') {
-        const dayKeys = ['sunday', 'monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday'];
-        const dayNames = dayOfWeek.split(',').map(d => {
-            const dayKey = dayKeys[parseInt(d.trim(), 10)] ?? 'sunday';
-            return t(`automations.days.${dayKey}`);
-        });
-        const dayStr = dayNames.length > 1
-            ? dayNames.slice(0, -1).join(', ') + ' ' + t('automations.and') + ' ' + dayNames[dayNames.length - 1]
-            : dayNames[0];
-        return t('automations.weeklyOn', { day: dayStr, time: timeStr });
+        return t('automations.weeklyOn', { day: formatDayOfWeek(dayOfWeek, t), time: timeStr });
     }
-
-    // Monthly
     if (dayOfMonth !== '*') {
-        const dom = parseInt(dayOfMonth, 10);
-        let dayStr = `${dom}`;
-        if (i18n.language.startsWith('en')) {
-            const suffixes = ['th', 'st', 'nd', 'rd'];
-            const v = dom % 100;
-            dayStr = `${dom}${suffixes[(v >= 11 && v <= 13) ? 0 : Math.min(v % 10, 4) > 3 ? 0 : v % 10]}`;
-        }
-        return t('automations.monthlyOn', { day: dayStr, time: timeStr });
+        return t('automations.monthlyOn', { day: formatDayOfMonth(parseInt(dayOfMonth, 10)), time: timeStr });
     }
 
     return config.schedule;
-}
-
-// Format next scheduled time
-function formatNextRun(nextScheduledAt: string | undefined, t: TFunction): string | null {
-    if (!nextScheduledAt) return null;
-
-    const next = new Date(nextScheduledAt);
-    const now = new Date();
-    const diffMs = next.getTime() - now.getTime();
-
-    if (diffMs < 0) return t('automations.overdue');
-
-    const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-    const diffDays = Math.floor(diffHours / 24);
-
-    if (diffDays > 0) {
-        return t('automations.inDays', { count: diffDays });
-    }
-    if (diffHours > 0) {
-        return t('automations.inHours', { count: diffHours });
-    }
-
-    const diffMinutes = Math.floor(diffMs / (1000 * 60));
-    if (diffMinutes > 0) {
-        return t('automations.inMinutes', { count: diffMinutes });
-    }
-
-    return t('automations.soon');
 }
 
 const STATUS_KEYS: Record<string, string> = {
